Create new branch only if it doesn't exist, otherwise it errors
The git_commit agent task status would be true if everything
goes well the first time but on a repeated request it would raise
an exception that the branch already exists and the task_manager
would set the task status false as it saw an exception. So the
admin would get a report with git_commit agent status set to
false even though it ran as expected the first time.

+++ git_commit.py
@@ -47,34 +47,35 @@ def git_commit(ch, method, properties, body):
         git.checkout('master')
         logger.debug('git pull')
         git.pull()
-        logger.debug('git checkout -b %s', branch_name)
-        git.checkout('-b', branch_name)
+        branch_exists = git.branch('--list', branch_name)
+        if not branch_exists:
+            logger.debug('git checkout -b %s', branch_name)
+            git.checkout('-b', branch_name)
+            logger.debug("open(%s, 'w'), open(%s, 'w')", user_ldif, group_ldif)
+            with open(user_ldif, 'w') as ldif_u,\
+                open(group_ldif, 'w') as ldif_g:
+                logger.debug(f"ldapsearch -LLL -x -H ldaps://ldapserver -b 'dc=cm,dc=cluster' uid={username} > {user_ldif}")
+                ldapsearch('-LLL', '-x', '-H', 'ldaps://ldapserver', '-b', "dc=cm,dc=cluster", f"uid={username}", _out=ldif_u)
+                logger.debug(f"ldapsearch -LLL -x -H ldapserver -b 'ou=Group,dc=cm,dc=cluster' cn={username} > {group_ldif}")
+                ldapsearch('-LLL', '-x', '-H', 'ldaps://ldapserver', '-b', "ou=Group,dc=cm,dc=cluster", f"cn={username}", _out=ldif_g)
+            logger.info('user ldif files generated.')
 
-        logger.debug("open(%s, 'w'), open(%s, 'w')", user_ldif, group_ldif)
-        with open(user_ldif, 'w') as ldif_u,\
-            open(group_ldif, 'w') as ldif_g:
-            logger.debug(f"ldapsearch -LLL -x -H ldaps://ldapserver -b 'dc=cm,dc=cluster' uid={username} > {user_ldif}")
-            ldapsearch('-LLL', '-x', '-H', 'ldaps://ldapserver', '-b', "dc=cm,dc=cluster", f"uid={username}", _out=ldif_u)
-            logger.debug(f"ldapsearch -LLL -x -H ldapserver -b 'ou=Group,dc=cm,dc=cluster' cn={username} > {group_ldif}")
-            ldapsearch('-LLL', '-x', '-H', 'ldaps://ldapserver', '-b', "ou=Group,dc=cm,dc=cluster", f"cn={username}", _out=ldif_g)
-        logger.info('user ldif files generated.')
+            logger.debug('git add %s', user_ldif)
+            git.add(user_ldif)
+            logger.debug('git add %s', group_ldif)
+            git.add(group_ldif)
+            logger.debug("git commit -m 'Added new cheaha user: %s'", username)
+            git.commit(m="Added new cheaha user: " + username)
+            logger.debug('git checkout master')
+            git.checkout('master')
 
-        logger.debug('git add %s', user_ldif)
-        git.add(user_ldif)
-        logger.debug('git add %s', group_ldif)
-        git.add(group_ldif)
-        logger.debug("git commit -m 'Added new cheaha user: %s'", username)
-        git.commit(m="Added new cheaha user: " + username)
-        logger.debug('git checkout master')
-        git.checkout('master')
+            logger.debug('git merge %s --no-ff --no-edit', branch_name)
+            git.merge(branch_name, '--no-ff', '--no-edit')
+            logger.debug('git push origin master')
+            git.push('origin', 'master')
+            # merge with gitlab api
 
-        logger.debug('git merge %s --no-ff --no-edit', branch_name)
-        git.merge(branch_name, '--no-ff', '--no-edit')
-        logger.debug('git push origin master')
-        git.push('origin', 'master')
-        # merge with gitlab api
-
-        logger.info('Added ldif files and committed to git repo')
+            logger.info('Added ldif files and committed to git repo')
 
         msg['success'] = True
     except Exception as exception:
